Fix segment info for ShadowEngine, remove test nocommit
